+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Additional Calypso requirements related to Stored Value operations (issue [#59]).
 - Additional Calypso requirements related to abortion of a secure session (issue [#57]).
 - Additional Calypso requirement related to the analysis of the APDU response length (issue [#62]).
+- Additional Calypso requirement related to the card extended mode command management (issue [#64]).
 
 ## [2.1.0] - 2022-02-01
 ### Added
@@ -61,6 +62,7 @@ It also brings many major API changes.
 [2.0.1]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/eclipse/keyple-card-calypso-java-lib/releases/tag/2.0.0
 
+[#64]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/64
 [#62]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/62
 [#59]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/59
 [#57]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/57